Fix build script
Apparently, the Google Fonts devs reserve the right to make breaking changes to their fonts without notice. Thus, the FontSource API can suffer malformed metadata and is reliant on Google Fonts devs to address it upstream.

This fixes the build script to compensate for missing weights/styles of the Robot font's Latin subsets. The cached metadata for the Roboto fonts is updated accordingly when errors occur downloading the fonts' ttf files.